requirements.txt
absl-py==0.8.1
addict==2.2.1
apex==0.1
astunparse==1.6.3
av==6.2.0
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2019.9.11
cffi==1.12.3
chainer==6.4.0
chainercv==0.13.1
chardet==3.0.4
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.0
dask==2.6.0
decorator==4.4.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
GDAL==3.1.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.8
imagecodecs==2019.5.22
imagecorruptions==1.1.0
imageio==2.6.1
joblib==0.14.1
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lxml==4.5.1
Markdown==3.1.1
matplotlib==3.1.1
mkl-service==2.3.0
mmcv==1.1.2
networkx==2.4
numpy==1.17.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.3.0
pandas==1.0.3
Pillow==6.1.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.2.0
pycocotools==2.0
pycparser==2.19
pydensecrf==1.0rc3
pyparsing==2.4.2
pyproj==2.6.1.post1
python-dateutil==2.8.0
pytz==2019.3
PyWavelets==1.0.3
pyyaml==5.1.2
regex==2020.7.14
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.16.1
scikit-learn==0.22.1
scipy==1.2.0
seaborn==0.10.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
tifffile==2019.7.26
toolz==0.10.0
torch==1.5.0+cu101
torchvision==0.6.0+cu101
tornado==6.0.3
tqdm==4.42.1
typing==3.6.6
typing-extensions==3.6.6
urllib3==1.25.6
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yapf==0.30.0
